cleanup up current path tracer logic
