+++ crates/appearance-path-tracer/Cargo.toml
@@ -8,6 +8,14 @@ publish.workspace = true
 version = "0.1.0"
 
 [dependencies]
+appearance-asset-database.workspace = true
+appearance-camera.workspace = true
+appearance-model.workspace = true
 appearance-profiling.workspace = true
+appearance-render-loop.workspace = true
+appearance-world.workspace = true
 
 glam.workspace = true
+log.workspace = true
+tinybvh.workspace = true
+uuid.workspace = true
